Create RQ1TidiedDataBin.csv
Data on the countries including their region and total waste. This will be used for my first visualisation. BinVar is just used as a variable that I've added so that when i reference the '?', any transformations or implementations i do will be for ALL the countries - like changing the point mark circles to the bin icons for all the countries